fix(dao nested pages): fix mobile view for nfts, settings and polls pages
