+++ astro_2.0/features/pages/nestedDaoPagesContent/SettingsPageContent/index.ts
@@ -1,0 +1,1 @@
+export * from './SettingsPageContent';
