+++ pages/dao/[dao]/tasks/polls/PollsPage.tsx
@@ -1,7 +1,4 @@
-import React, { FC } from 'react';
-
-// Constants
-import { ALL_DAOS_URL } from 'constants/routing';
+import React, { FC, useMemo } from 'react';
 
 // Types
 import { PaginationResponse } from 'types/api';
@@ -14,17 +11,12 @@ import {
 
 // Components
 import { Feed } from 'astro_2.0/features/Feed';
-import { NavLink } from 'astro_2.0/components/NavLink';
-import { BreadCrumbs } from 'astro_2.0/components/BreadCrumbs';
-import { PolicyAffectedWarning } from 'astro_2.0/components/PolicyAffectedWarning';
-import { DaoDetailsMinimized } from 'astro_2.0/components/DaoDetails';
-import { useCreateProposal } from 'astro_2.0/features/CreateProposal/hooks';
+import { NestedDaoPageWrapper } from 'astro_2.0/features/pages/nestedDaoPagesContent/NestedDaoPageWrapper';
 
 // Hooks
-import { useDaoCustomTokens } from 'hooks/useCustomTokens';
+import { useGetBreadcrumbsConfig } from 'hooks/useGetBreadcrumbsConfig';
 
 import { DaoContext } from 'types/context';
-import { NoResultsView } from 'astro_2.0/components/NoResultsView';
 import styles from './Polls.module.scss';
 
 export interface PollsPageProps {
@@ -34,61 +26,40 @@ export interface PollsPageProps {
 }
 
 const PollsPage: FC<PollsPageProps> = ({
-  daoContext: { dao, policyAffectsProposals, userPermissions },
+  daoContext,
   initialPollsData,
   initialProposalsStatusFilterValue,
 }) => {
-  const { tokens } = useDaoCustomTokens();
+  const breadcrumbsConfig = useGetBreadcrumbsConfig(
+    daoContext.dao.id,
+    daoContext.dao.displayName
+  );
 
-  const [CreateProposal, toggleCreateProposal] = useCreateProposal();
+  const breadcrumbs = useMemo(() => {
+    return [
+      breadcrumbsConfig.ALL_DAOS_URL,
+      breadcrumbsConfig.SINGLE_DAO_PAGE,
+      breadcrumbsConfig.POLLS,
+    ];
+  }, [breadcrumbsConfig]);
 
   return (
-    <div className={styles.root}>
-      <BreadCrumbs className={styles.breadcrumbs}>
-        <NavLink href={ALL_DAOS_URL}>All DAOs</NavLink>
-        <NavLink href={`/dao/${dao.id}`}>{dao?.displayName || dao?.id}</NavLink>
-        <span>Polls</span>
-      </BreadCrumbs>
-
-      <div className={styles.dao}>
-        <DaoDetailsMinimized
-          dao={dao}
-          userPermissions={userPermissions}
-          onCreateProposalClick={toggleCreateProposal}
-        />
-
-        <CreateProposal
-          className={styles.createProposal}
-          dao={dao}
-          userPermissions={userPermissions}
-          key={Object.keys(tokens).length}
-          daoTokens={tokens}
-          proposalVariant={ProposalVariant.ProposePoll}
-          showFlag={false}
-          onClose={toggleCreateProposal}
-        />
-
-        <PolicyAffectedWarning
-          data={policyAffectsProposals}
-          className={styles.warningWrapper}
-        />
-      </div>
-
-      {initialPollsData.count === 0 ? (
-        <NoResultsView title="No polls available" />
-      ) : (
-        <Feed
-          title={<h1 className={styles.header}>Polls</h1>}
-          dao={dao}
-          showFlag={false}
-          className={styles.feed}
-          category={ProposalCategories.Polls}
-          initialProposals={initialPollsData}
-          headerClassName={styles.feedHeader}
-          initialProposalsStatusFilterValue={initialProposalsStatusFilterValue}
-        />
-      )}
-    </div>
+    <NestedDaoPageWrapper
+      daoContext={daoContext}
+      breadcrumbs={breadcrumbs}
+      defaultProposalType={ProposalVariant.ProposeCreateBounty}
+    >
+      <Feed
+        title={<h1 className={styles.header}>Polls</h1>}
+        dao={daoContext.dao}
+        showFlag={false}
+        className={styles.feed}
+        category={ProposalCategories.Polls}
+        initialProposals={initialPollsData}
+        headerClassName={styles.feedHeader}
+        initialProposalsStatusFilterValue={initialProposalsStatusFilterValue}
+      />
+    </NestedDaoPageWrapper>
   );
 };
